renamed:    nanopub/nanopub.py -> nanopub/models.py
