+++ nanopub/__init__.py
@@ -1,3 +1,3 @@
-from .nanopub import Nanopub
+from .models import Nanopub
 from .client import NanopubClient
 from ._version import __version__
